First cut at an OSC output.
It's still pretty rough.  Please read the Roadmap files
at the top level of the tree and in the plg_osc directory.

I've tried to provide some explanatory notes on how this
agent works, and I've done a bit of a brain dump to outline
some of the broader eigend concepts.

We'll move this kind of thing to the Wiki eventually.