Reworked lights.
Light buffers now hold (packed) tuples of (row,column,value)
instead of a simple list of values.  row and column obey all
the relative rules used elsewhere.  When the row is 0, the
column is an absolute key number.

This allows (for instance) relative talkers to be used from
multiple kgroup outputs and still get the lighting right.