Per agent lexicons.
There are 2 ways of adding new words.  The simplest way is to add
a file to your agent tree called (eg) 'lexicon.txt'  This has the
same format as the file in pibelcanto/

Then add that file to your PiAgent line in your SConscript.

    env.PiAgent('strummer','plg_simple','strummer_plg',cversion='1.0.0',lexicon='lexicon.txt')

This adds the words into the Manifest file.  It is picked up from
there by EigenD and exported on your agents behalf.  This is necessary
if you wish to coin a new word for the name of your agent; the vocabulary
needs to be in place before the agent is loaded.

The other way is to supply a vocab dictionary to an Atom (or Agent)
A lexicon dictionary has the form:

    my_lex = { 'english': [ '123', 'verb' ] }

you can call the set_vocab method on Atom/Agent or pass a vocab
keyword to the constructor.  This would allow new words to be
coined on-the-fly.

The dictionary panel in commander should update itself as the vocabulary
changes.