Automate code locking.
Windows and Mac implementations are completely different.

Windows just uses linker attributes on various sections.  These are set by the
build system when a shared library or pip binding is tagged as locked.

Mac OS needs a handler running which reacts to dll loading and manually mlocks
the relevant sections at load time.  For this, the build system adds a marker
file with a special symbol in it, which the run time support looks for.