Made the alpha (and tau) debouncing system configurable.
The default is 20000 us.

This seems to cause missing keypresses quite easily with
stickier keys (ie, coeff of friction between finger and key
is high)  The same effect seems to happen on other
keyboards, to greater or lesser extent.

If you are stefan, setting this to 25000 or higher should
help.  We could no longer reproduce the problem with stefans
keyboard with this setting.

The maximum is 31500 (31.5ms)

keyboard 1 debounce to 25000 set