Added support for changing the light map by using a bitmap string instead of a list or coordinates and colours.
Use a PUT method on the /physical or /musical paths with the bitmap and
the agent will detect whether to parse it as a bitmap or a list.

The bitmap has rows and courses as lines, columns and keys as characters
in the lines, and the letters 'R', 'G', 'O' for the colours red, green
and orange. A space is used to leave a light off.

For instance:

RR GG OO
  O  R  G

With curl and bash, this can be set in one go like this:

curl http://localhost:1024/physical -X PUT --data $'RR GG OO\n  O  R  G'