Re-worked noun scoping.  Rigs are no longer special;
'all scaler' will match all scalers in the system.

There are various heuristics to deal with ambiguity
between (say) scaler 1 at the top level and scaler 1
inside a rig.

Extended the identify verb to show the full name
(including rig) of an agent.

Moved the rig create/un create verbs into EigenD, so
to create an agent in a rig you now say:

    summer in rig 1 create

to eigend, instead of talking to the rig.

The net result of all this is that:

    illuminator create it to chord illuminator name ify

should now work again.